+++ x/gov/client/cli/prompt.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"reflect" // #nosec
+	"reflect"
 	"sort"
 	"strconv"
 	"strings"
@@ -12,14 +12,12 @@ import (
 	"github.com/manifoldco/promptui"
 	"github.com/spf13/cobra"
 
-	"cosmossdk.io/core/address"
-	"cosmossdk.io/x/gov/types"
-
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	"github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
 const (
@@ -29,43 +27,21 @@ const (
 	draftMetadataFileName = "draft_metadata.json"
 )
 
-var suggestedProposalTypes = []proposalType{
-	{
-		Name:    proposalText,
-		MsgType: "", // no message for text proposal
-	},
-	{
-		Name:    "community-pool-spend",
-		MsgType: "/cosmos.protocolpool.v1.MsgCommunityPoolSpend",
-	},
-	{
-		Name:    "software-upgrade",
-		MsgType: "/cosmos.upgrade.v1beta1.MsgSoftwareUpgrade",
-	},
-	{
-		Name:    "cancel-software-upgrade",
-		MsgType: "/cosmos.upgrade.v1beta1.MsgCancelUpgrade",
-	},
-	{
-		Name:    "submit-budget-proposal",
-		MsgType: "/cosmos.protocolpool.v1.MsgSubmitBudgetProposal",
-	},
-	{
-		Name:    "create-continuous-fund",
-		MsgType: "/cosmos.protocolpool.v1.MsgCreateContinuousFund",
-	},
-	{
-		Name:    proposalOther,
-		MsgType: "", // user will input the message type
-	},
+// ProposalMetadata is the metadata of a proposal
+// This metadata is supposed to live off-chain when submitted in a proposal
+type ProposalMetadata struct {
+	Title             string `json:"title"`
+	Authors           string `json:"authors"`
+	Summary           string `json:"summary"`
+	Details           string `json:"details"`
+	ProposalForumUrl  string `json:"proposal_forum_url"` // named 'Url' instead of 'URL' for avoiding the camel case split
+	VoteOptionContext string `json:"vote_option_context"`
 }
 
 // Prompt prompts the user for all values of the given type.
 // data is the struct to be filled
-// namePrefix is the name to be displayed as "Enter <namePrefix> <field>"
-// TODO: when bringing this in autocli, use proto message instead
-// this will simplify the get address logic
-func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, error) {
+// namePrefix is the name to be display as "Enter <namePrefix> <field>"
+func Prompt[T any](data T, namePrefix string) (T, error) {
 	v := reflect.ValueOf(&data).Elem()
 	if v.Kind() == reflect.Interface {
 		v = reflect.ValueOf(data)
@@ -75,15 +51,10 @@ func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, er
 	}
 
 	for i := 0; i < v.NumField(); i++ {
-		// if the field is a struct skip or not slice of string or int then skip
-		switch v.Field(i).Kind() {
-		case reflect.Struct:
-			// TODO(@julienrbrt) in the future we can add a recursive call to Prompt
+		if v.Field(i).Kind() == reflect.Struct || v.Field(i).Kind() == reflect.Slice {
+			// if the field is a struct skip
+			// in a future we can add a recursive call to Prompt
 			continue
-		case reflect.Slice:
-			if v.Field(i).Type().Elem().Kind() != reflect.String && v.Field(i).Type().Elem().Kind() != reflect.Int {
-				continue
-			}
 		}
 
 		// create prompts
@@ -93,18 +64,17 @@ func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, er
 		}
 
 		fieldName := strings.ToLower(v.Type().Field(i).Name)
+		// validation per field name
+		if strings.Contains(fieldName, "url") {
+			prompt.Validate = client.ValidatePromptURL
+		}
 
 		if strings.EqualFold(fieldName, "authority") {
 			// pre-fill with gov address
-			defaultAddr, err := addressCodec.BytesToString(authtypes.NewModuleAddress(types.ModuleName))
-			if err != nil {
-				return data, err
-			}
-			prompt.Default = defaultAddr
+			prompt.Default = authtypes.NewModuleAddress(types.ModuleName).String()
 			prompt.Validate = client.ValidatePromptAddress
 		}
 
-		// TODO(@julienrbrt) use scalar annotation instead of dumb string name matching
 		if strings.Contains(fieldName, "addr") ||
 			strings.Contains(fieldName, "sender") ||
 			strings.Contains(fieldName, "voter") ||
@@ -124,31 +94,11 @@ func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, er
 		case reflect.String:
 			v.Field(i).SetString(result)
 		case reflect.Int:
-			resultInt, err := strconv.ParseInt(result, 10, 0)
-			if err != nil {
-				return data, fmt.Errorf("invalid value for int: %w", err)
-			}
-			// If a value was successfully parsed the ranges of:
-			//      [minInt,     maxInt]
-			// are within the ranges of:
-			//      [minInt64, maxInt64]
-			// of which on 64-bit machines, which are most common,
-			// int==int64
-			v.Field(i).SetInt(resultInt)
-		case reflect.Slice:
-			switch v.Field(i).Type().Elem().Kind() {
-			case reflect.String:
-				v.Field(i).Set(reflect.ValueOf([]string{result}))
-			case reflect.Int:
-				resultInt, err := strconv.ParseInt(result, 10, 0)
-				if err != nil {
-					return data, fmt.Errorf("invalid value for int: %w", err)
-				}
-
-				v.Field(i).Set(reflect.ValueOf([]int{int(resultInt)}))
-			}
+			resultInt, _ := strconv.Atoi(result)
+			v.Field(i).SetInt(int64(resultInt))
 		default:
-			// skip any other types
+			// skip other types
+			// possibly in the future we can add more types (like slices)
 			continue
 		}
 	}
@@ -156,24 +106,22 @@ func Prompt[T any](data T, namePrefix string, addressCodec address.Codec) (T, er
 	return data, nil
 }
 
-type proposalType struct {
-	Name    string
+type proposalTypes struct {
+	Type    string
 	MsgType string
 	Msg     sdk.Msg
 }
 
 // Prompt the proposal type values and return the proposal and its metadata
-func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec address.Codec) (*proposal, types.ProposalMetadata, error) {
-	metadata, err := PromptMetadata(skipMetadata, addressCodec)
+func (p *proposalTypes) Prompt(cdc codec.Codec) (*proposal, ProposalMetadata, error) {
+	proposal := &proposal{}
+
+	// set metadata
+	metadata, err := Prompt(ProposalMetadata{}, "proposal")
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to set proposal metadata: %w", err)
 	}
-
-	proposal := &proposal{
-		Metadata: "ipfs://CID", // the metadata must be saved on IPFS, set placeholder
-		Title:    metadata.Title,
-		Summary:  metadata.Summary,
-	}
+	proposal.Metadata = "ipfs://CID"
 
 	// set deposit
 	depositPrompt := promptui.Prompt{
@@ -190,7 +138,7 @@ func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec a
 	}
 
 	// set messages field
-	result, err := Prompt(p.Msg, "msg", addressCodec)
+	result, err := Prompt(p.Msg, "msg")
 	if err != nil {
 		return nil, metadata, fmt.Errorf("failed to set proposal message: %w", err)
 	}
@@ -200,61 +148,63 @@ func (p *proposalType) Prompt(cdc codec.Codec, skipMetadata bool, addressCodec a
 		return nil, metadata, fmt.Errorf("failed to marshal proposal message: %w", err)
 	}
 	proposal.Messages = append(proposal.Messages, message)
-
 	return proposal, metadata, nil
 }
 
-// getProposalSuggestions suggests a list of proposal types
-func getProposalSuggestions() []string {
-	types := make([]string, len(suggestedProposalTypes))
-	for i, p := range suggestedProposalTypes {
-		types[i] = p.Name
+var supportedProposalTypes = []proposalTypes{
+	{
+		Type:    proposalText,
+		MsgType: "", // no message for text proposal
+	},
+	{
+		Type:    "community-pool-spend",
+		MsgType: "/cosmos.distribution.v1beta1.MsgCommunityPoolSpend",
+	},
+	{
+		Type:    "software-upgrade",
+		MsgType: "/cosmos.upgrade.v1beta1.MsgSoftwareUpgrade",
+	},
+	{
+		Type:    "cancel-software-upgrade",
+		MsgType: "/cosmos.upgrade.v1beta1.MsgCancelUpgrade",
+	},
+	{
+		Type:    proposalOther,
+		MsgType: "", // user will input the message type
+	},
+}
+
+func getProposalTypes() []string {
+	types := make([]string, len(supportedProposalTypes))
+	for i, p := range supportedProposalTypes {
+		types[i] = p.Type
 	}
 	return types
 }
 
-// PromptMetadata prompts for proposal metadata or only title and summary if skip is true
-func PromptMetadata(skip bool, addressCodec address.Codec) (types.ProposalMetadata, error) {
-	if !skip {
-		metadata, err := Prompt(types.ProposalMetadata{}, "proposal", addressCodec)
-		if err != nil {
-			return metadata, fmt.Errorf("failed to set proposal metadata: %w", err)
-		}
-
-		return metadata, nil
-	}
-
-	// prompt for title and summary
-	titlePrompt := promptui.Prompt{
-		Label:    "Enter proposal title",
-		Validate: client.ValidatePromptNotEmpty,
-	}
-
-	title, err := titlePrompt.Run()
+func getProposalMsg(cdc codec.Codec, input string) (sdk.Msg, error) {
+	var msg sdk.Msg
+	bz, err := json.Marshal(struct {
+		Type string `json:"@type"`
+	}{
+		Type: input,
+	})
 	if err != nil {
-		return types.ProposalMetadata{}, fmt.Errorf("failed to set proposal title: %w", err)
+		return nil, err
 	}
 
-	summaryPrompt := promptui.Prompt{
-		Label:    "Enter proposal summary",
-		Validate: client.ValidatePromptNotEmpty,
+	if err := cdc.UnmarshalInterfaceJSON(bz, &msg); err != nil {
+		return nil, fmt.Errorf("failed to determined sdk.Msg from %s proposal type : %w", input, err)
 	}
 
-	summary, err := summaryPrompt.Run()
-	if err != nil {
-		return types.ProposalMetadata{}, fmt.Errorf("failed to set proposal summary: %w", err)
-	}
-
-	return types.ProposalMetadata{Title: title, Summary: summary}, nil
+	return msg, nil
 }
 
 // NewCmdDraftProposal let a user generate a draft proposal.
 func NewCmdDraftProposal() *cobra.Command {
-	flagSkipMetadata := "skip-metadata"
-
 	cmd := &cobra.Command{
 		Use:          "draft-proposal",
-		Short:        "Generate a draft proposal json file. The generated proposal json contains only one message (skeleton).",
+		Short:        "Generate a draft proposal json file. The generated proposal json contains only one message.",
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -265,24 +215,24 @@ func NewCmdDraftProposal() *cobra.Command {
 			// prompt proposal type
 			proposalTypesPrompt := promptui.Select{
 				Label: "Select proposal type",
-				Items: getProposalSuggestions(),
+				Items: getProposalTypes(),
 			}
 
-			_, selectedProposalType, err := proposalTypesPrompt.Run()
+			_, proposalType, err := proposalTypesPrompt.Run()
 			if err != nil {
 				return fmt.Errorf("failed to prompt proposal types: %w", err)
 			}
 
-			var proposal proposalType
-			for _, p := range suggestedProposalTypes {
-				if strings.EqualFold(p.Name, selectedProposalType) {
+			var proposal proposalTypes
+			for _, p := range supportedProposalTypes {
+				if strings.EqualFold(p.Type, proposalType) {
 					proposal = p
 					break
 				}
 			}
 
 			// create any proposal type
-			if proposal.Name == proposalOther {
+			if proposal.Type == proposalOther {
 				// prompt proposal type
 				msgPrompt := promptui.Select{
 					Label: "Select proposal message type:",
@@ -302,43 +252,37 @@ func NewCmdDraftProposal() *cobra.Command {
 			}
 
 			if proposal.MsgType != "" {
-				proposal.Msg, err = sdk.GetMsgFromTypeURL(clientCtx.Codec, proposal.MsgType)
+				proposal.Msg, err = getProposalMsg(clientCtx.Codec, proposal.MsgType)
 				if err != nil {
 					// should never happen
 					panic(err)
 				}
 			}
 
-			skipMetadataPrompt, _ := cmd.Flags().GetBool(flagSkipMetadata)
-
-			result, metadata, err := proposal.Prompt(clientCtx.Codec, skipMetadataPrompt, clientCtx.AddressCodec)
+			prop, metadata, err := proposal.Prompt(clientCtx.Codec)
 			if err != nil {
 				return err
 			}
 
-			if err := writeFile(draftProposalFileName, result); err != nil {
+			if err := writeFile(draftMetadataFileName, metadata); err != nil {
 				return err
 			}
 
-			if !skipMetadataPrompt {
-				if err := writeFile(draftMetadataFileName, metadata); err != nil {
-					return err
-				}
+			if err := writeFile(draftProposalFileName, prop); err != nil {
+				return err
 			}
 
-			cmd.Println("The draft proposal has successfully been generated.\nProposals should contain off-chain metadata, please upload the metadata JSON to IPFS.\nThen, replace the generated metadata field with the IPFS CID.")
+			fmt.Printf("Your draft proposal has successfully been generated.\nProposals should contain off-chain metadata, please upload the metadata JSON to IPFS.\nThen, replace the generated metadata field with the IPFS CID.\n")
 
 			return nil
 		},
 	}
 
 	flags.AddTxFlagsToCmd(cmd)
-	cmd.Flags().Bool(flagSkipMetadata, false, "skip metadata prompt")
 
 	return cmd
 }
 
-// writeFile writes the input to the file
 func writeFile(fileName string, input any) error {
 	raw, err := json.MarshalIndent(input, "", " ")
 	if err != nil {